YOLO v8n Model
In this project, the yolov8n model in used on the traffic flow video. The main purpose was to identify the flow of traffic. In addition, interaction between walking people and vehicles on the road was also monitored. There are walking lines on the road, so it was also identified how many people are using that. 
In a nutshell, it was sort of investigative project in which YOLO v8n is used to identify the real time images detection.